cleanup files and improve logging
