+++ python/generate_website.py
@@ -1,120 +1,71 @@
 import get_sheet_data
+import get_drive_data
+import update_json_and_sheet
+import config
 import generate_css_html
 import generate_other_html
 import generate_overflow_html
 import platform
 import subprocess
-import test_mapname_filename_match
-import test_drive_matches_sheet
-import write_screenshots_to_sheet
 import re
 
-TESTS_ENABLED = True
-WRITE_TO_SHEET = True
+SHEET_DATA = config.get_post_processed_sheet_data_from_json()
+MAP_NAME_INDEX = config.get_map_name_index()
+MAP_DOWNLOAD_INDEX = config.get_map_download_index()
+SCREENSHOT_INDEX = config.get_screenshot_index()
+GAME_INDEX = config.get_game_index()
+OVERFLOW_INDEX = config.get_overflow_index()
 
-def main():
-    data = get_sheet_data.get_data()
-    collapsible_with_dl, collapsible_no_dl, overflow = create_collapsible(data)
-    css_dl, css_no_dl, other_dl, other_no_dl = split_map_by_category(collapsible_with_dl, collapsible_no_dl)
-    generate_css_html.build(css_dl, css_no_dl)
-    generate_other_html.build(other_dl, other_no_dl)
-    generate_overflow_html.build(overflow)
+def create_collapsibles(data):
+    collapsibles = {
+        "css_dl": [],
+        "css_no_dl": [],
+        "other_dl": [],
+        "other_no_dl": [],
+        "overflow": []
+    }
 
-def create_collapsible(data):
-    collapsible_list_dl = []
-    collapsible_list_no_dl = []
-    overflow_list = []
-
-    # i dont really like this, maybe a better way in the future?
-    map_name_index = 0
-    link_index = 6
-    img_index = 7
-    overflow_index = 8
-
-    for row in data[1:]:  # Skip the header row
-        content = []
-        has_dl = False
-        overflow = False
-
-        # this could be done by setting variables to each index without a second for loop and the if statements
-        # but this way feels a little more readable to me
-        
-        for index, item in enumerate(row[0:], start=0): 
-            column_header = data[0][index]
-
-            if index == map_name_index:
-                map_name = item
-
-            elif index == overflow_index and item is "yes":
-                overflow = True
-
-            elif index == link_index and "drive.google.com" in item and overflow == False:
-                # Handle links for the specified index
-                drive_link_from_sheet = item
-                site_link = f'<a href="{drive_link_from_sheet}">{map_name}.zip</a>' # don't feel like calling drive api every time to get file type.  it's always zip for downloads
-                content.append(f'<b>{column_header}:</b><br />&emsp;{site_link}') 
-                has_dl = True
-            
-            elif index == link_index and "drive.google.com" not in item and overflow == False:
-                has_dl = False
-
-            elif index == img_index and "drive.google.com" in item:
-                image_url = item # just for readability
-                lienuc_image_url = build_lien_uc(image_url)
-
-                # Handle images with missing or error alt text for the specified index
-                img_alt = f'{map_name}'
-                img_link = f'<img {lienuc_image_url} alt="{img_alt}" class="ImgThumbnail" loading="lazy">'
-                content.append(f'<b>screenshot:</b><br />&emsp;{img_link}')
-
-                # site link always comes last, which is after screenshot
-                site_link = (f'<b>site link:</b><br />&emsp;<a href="#{map_name}">{map_name}</a>')
-                content.append(site_link)
-
-            elif index == overflow_index and item is not None:
-                pass
-
-            else:
-                # Handle regular content
-                content.append(f'<b>{column_header}:</b><br />&emsp;{item}') # noted above, data[0] = column header name
-
+    column_header = data[0]  # Assuming data[0] contains column headers
+    
+    for row in data[1:]:
+        # ok what the hell is this line?
+        # one big boy list comprehension
+        # it sets the content, which is one line of HTML
+        # equal to the header in cell 
+        # which is the pair found in each item of the list from the zip of column_header and row
+        # the zip makes it so each column header is always paired with its cell
+        # for reference when building the content for a collapsible.
+        # it also leaves out the overflow, because overflow column is only used for sorting lists
+        content = [f'<b>{header}:</b><br />&emsp;{cell}' for header, cell in zip(column_header, row) if column_header.index(header) != OVERFLOW_INDEX]
         content_html = '<br />'.join(content)
-        collapsible_html = f'\n<div id="{map_name}">\n\t<button type="button" class="collapsible">{map_name}</button>\n\t<div id="{map_name}" class="content"><p>{content_html}</p></div>\n</div>'
-        # This line is a little absurd
-        # JS found in collapsible_handler.js utilizes the nested divs with the same name
-        # if a jump link is followed to a matching map name, the JS expands that collapsible.
+        collapsible_html = f'\n<div id="{row[MAP_NAME_INDEX]}">\n\t<button type="button" class="collapsible">{row[MAP_NAME_INDEX]}</button>\n\t<div id="{row[MAP_NAME_INDEX]}" class="content"><p>{content_html}</p></div>\n</div>'
+        
+        has_dl = "drive.google.com" in row[MAP_DOWNLOAD_INDEX]
+        css = "CSS" in row[GAME_INDEX]
+        overflow = "x" in row[OVERFLOW_INDEX]
 
-        if has_dl and not overflow:
-            collapsible_list_dl.append(collapsible_html)
+        category = None
 
-        elif not has_dl and not overflow:
-            collapsible_list_no_dl.append(collapsible_html)
-
+        # now that a collapsible element has been built for every map,
+        # we need to split them out into categories
+        # so they can be placed on the correct column of the correct page.
+        if has_dl and css and not overflow:
+            category = "css_dl"
+        elif not has_dl and css and not overflow:
+            category = "css_no_dl"
+        elif has_dl and not css and not overflow:
+            category = "other_dl"
+        elif not has_dl and not css and not overflow:
+            category = "other_no_dl"
         elif overflow:
-            overflow_list.append(collapsible_html)
-
-    return collapsible_list_dl, collapsible_list_no_dl, overflow_list
-
-def split_map_by_category(collapsible_list_dl, collapsible_list_no_dl):
-    collapsible_list_css_dl = []
-    collapsible_list_css_no_dl = []
-    collapsible_list_1p6_dl = []
-    collapsible_list_1p6_no_dl = []
-
-    for item in range(len(collapsible_list_dl)):
-        if "CSS" in collapsible_list_dl[item]:
-            collapsible_list_css_dl.append(collapsible_list_dl[item])
-        else:
-            collapsible_list_1p6_dl.append(collapsible_list_dl[item])
+            category = "overflow"
+        
+        if category is not None:
+            collapsibles[category].append(collapsible_html)
     
-    for item in range(len(collapsible_list_no_dl)):
-        if "CSS" in collapsible_list_no_dl[item]:
-            collapsible_list_css_no_dl.append(collapsible_list_no_dl[item])
-        else:
-            collapsible_list_1p6_no_dl.append(collapsible_list_no_dl[item])
-    
-    return collapsible_list_css_dl, collapsible_list_css_no_dl, collapsible_list_1p6_dl, collapsible_list_1p6_no_dl
+    return tuple(collapsibles.values())
 
+# lint the html, I don't like this too much since it's platform dependent
 def tidy_html():
     # check if being run on linux, if so, is tidy installed? if so, run html tidy script
     if platform.system() == "Linux":
@@ -130,26 +81,37 @@ def tidy_html():
         except subprocess.CalledProcessError:
             print("Tidy is not installed. Please install it before running the HTML tidy script.")
 
-def build_lien_uc(item):
-    # https://lienuc.com/ - Jan 2024 google broke how embeds worked
-    # so, strip IDs from full URL (item) and run through lienuc
-    regex_for_id = re.compile(r'id=(.+)') # regex for everything after id= in the google url
-    image_id = regex_for_id.search(item).group(1)
-    lienuc_image_url = f'crossorigin="anonymous" src="http://drive.lienuc.com/uc?id={image_id}"'
-    return lienuc_image_url    
+if __name__ == "__main__":
+    print("==========GETTING ALL DATA FROM GOOGLE APIS==========")
+    print("") # yes i know i can do \n i just like how this breaks it up visually
 
-if __name__ == '__main__':
-    if WRITE_TO_SHEET == True:
-        print("Writing links to all screenshots in drive to sheet")
-        write_screenshots_to_sheet.generate_rows_with_screenshot()
-    main()
-    if TESTS_ENABLED == True:
-        print("Testing if all drive files have sheet entry")
-        test_drive_matches_sheet.get_downloads_for_missing_maps(WRITE_TO_SHEET) # this is confusing, maybe split into a test and a write
-        print("Testing if all mapnames match filenames for downloads")
-        test_mapname_filename_match.check_names()
-        print("all done with tests!!")
-    else:
-        print("no tests, all done")
-    print("tidying html")
-    tidy_html()
+    get_sheet_data.get_data() # generate preprocessed json file from API
+    get_drive_data.get_screenshot_items() # generate json for screenshots from api
+    get_drive_data.get_map_items() # generate json for maps from api
+    
+    print("")
+    print("==========COMBINING SHEET AND DRIVE JSONS==========")
+    print("")
+
+    update_json_and_sheet.pad_rows() # pad rows if there are empties, so it's always correct length
+    update_json_and_sheet.match_screenshots_and_downloads_to_sheet() # match sheet data against screenshot and map drive json's, also format links for html
+    update_json_and_sheet.add_jump_links() # build jump links for site
+
+    print("")
+    print("==========WRITING UPDATED JSON TO FILE, THEN TO SHEET==========")
+    print("")
+
+    update_json_and_sheet.write_processed_json_to_file() # write the processed json to a file
+    update_json_and_sheet.write_processed_json_to_sheet() # write the processed json to sheet
+    
+    print("")
+    print("==========BUILDING WEBSITE PAGES==========")
+    print("")
+
+    css_dl, css_no_dl, other_dl, other_no_dl, overflow = create_collapsibles(SHEET_DATA) # get variables for all collapsibles
+    generate_css_html.build(css_dl, css_no_dl) # generate css HTML page (index.html)
+    print("built index.html")
+    generate_other_html.build(other_dl, other_no_dl) # generate other maps HTML page (other.html)
+    print("built other.html")
+    generate_overflow_html.build(overflow) # generate overflow maps HTML page (overflow.html)
+    print("built overflow.html")
