--- python/json_as_list.py
@@ -1,6 +1,0 @@
-import json
-
-def load_file_data(file_name):
-    with open(file_name, 'r') as json_file:
-        sheet_data_as_list = json.load(json_file)
-    return sheet_data_as_list
